+++ frontend/src/pages/Users/DialogContainer.js
@@ -29,7 +29,11 @@ import PasswordDialog from "./PasswordDialog";
 
 class DialogContainer extends Component {
   componentDidUpdate(prevProps) {
-    if (!this.props.dashboardDialogShown && prevProps.dashboardDialogShown) {
+    if (
+      this.props.dialogType === "editUserPermissions" &&
+      !this.props.dashboardDialogShown &&
+      prevProps.dashboardDialogShown
+    ) {
       this.props.listGlobalPermissions();
     }
   }
@@ -38,6 +42,7 @@ class DialogContainer extends Component {
     return (
       <div>
         {this.props.dashboardDialogShown ? <Dialog {...this.props} /> : null}
+
         {this.props.passwordDialogShown ? (
           <PasswordDialog
             classes={this.props.classes}
@@ -62,7 +67,8 @@ const mapStateToProps = state => {
     dialogType: state.getIn(["users", "dialogType"]),
     editId: state.getIn(["users", "editId"]),
     userToAdd: state.getIn(["users", "userToAdd"]),
-    users: state.getIn(["login", "enabledUsers"]),
+    enabledUsers: state.getIn(["login", "enabledUsers"]),
+    disabledUsers: state.getIn(["login", "disabledUsers"]),
     loggedInUserId: state.getIn(["login", "id"]),
     groups: state.getIn(["users", "groups"]),
     groupToAdd: state.getIn(["users", "groupToAdd"]),